+++ src/deluge/playback/mode/session.cpp
@@ -86,7 +86,7 @@ const Colour defaultClipSectionColours[] = {RGB::fromHue(102), // bright light b
 
 Session::Session() {
 	cancelAllLaunchScheduling();
-	lastSectionArmed = 255;
+	lastSectionArmed = REALLY_OUT_OF_RANGE;
 }
 
 void Session::armAllClipsToStop(int32_t afterNumRepeats) {
@@ -137,18 +137,20 @@ void Session::armAllClipsToStop(int32_t afterNumRepeats) {
 }
 
 void Session::armNextSection(int32_t oldSection, int32_t numRepetitions) {
-	if (numRepetitions == -1) {
+	if (numRepetitions <= -1) {
 		numRepetitions = currentSong->sections[oldSection].numRepetitions;
 	}
 	if (currentSong->sessionLayout == SessionLayoutType::SessionLayoutTypeRows) {
 		if (currentSong->sessionClips.getClipAtIndex(0)->section != oldSection) {
 
+			int32_t newSection = SECTION_OUT_OF_RANGE;
 			for (int32_t c = 1; c < currentSong->sessionClips.getNumElements(); c++) { // NOTE: starts at 1, not 0
 				Clip* clip = currentSong->sessionClips.getClipAtIndex(c);
-
-				if (clip->section == oldSection) {
-					int32_t newSection =
-					    currentSong->sessionClips.getClipAtIndex(c - 1)->section; // Grab section from next Clip down
+				int32_t tempSection = currentSong->sessionClips.getClipAtIndex(c - 1)->section;
+				if (currentSong->sections[tempSection].numRepetitions != LAUNCH_EXCLUSIVE) {
+					newSection = tempSection;
+				}
+				if (clip->section == oldSection && newSection != SECTION_OUT_OF_RANGE) {
 					userWantsToArmClipsToStartOrSolo(newSection, nullptr, true, false, false, numRepetitions, false);
 					lastSectionArmed = newSection;
 					return;
@@ -159,15 +161,24 @@ void Session::armNextSection(int32_t oldSection, int32_t numRepetitions) {
 	// grid mode - just go to the next section, no need to worry about what order they're in
 	else if (currentSong->sessionLayout == SessionLayoutType::SessionLayoutTypeGrid) {
 		if (oldSection < kMaxNumSections) {
-			userWantsToArmClipsToStartOrSolo(oldSection + 1, nullptr, true, false, false, numRepetitions, false);
-			lastSectionArmed = oldSection + 1;
+			int32_t newSection = SECTION_OUT_OF_RANGE;
+			for (int32_t c = oldSection + 1; c <= kMaxNumSections; ++c) {
+				if (currentSong->sections[c].numRepetitions != LAUNCH_EXCLUSIVE) {
+					newSection = c;
+					break;
+				}
+			}
+			if (newSection != SECTION_OUT_OF_RANGE) {
+				userWantsToArmClipsToStartOrSolo(newSection, nullptr, true, false, false, numRepetitions, false);
+				lastSectionArmed = newSection;
+			}
 			return;
 		}
 	}
 
 	// If we're here, that was the last section
 	armAllClipsToStop(numRepetitions);
-	lastSectionArmed = 254;
+	lastSectionArmed = SECTION_OUT_OF_RANGE;
 }
 
 // Returns whether it began
@@ -520,7 +531,7 @@ stopOnlyIfOutputTaken:
 	// separate undoability
 	actionLogger.closeAction(ActionType::RECORD);
 
-	bool sectionWasJustLaunched = (lastSectionArmed < 254);
+	bool sectionWasJustLaunched = (lastSectionArmed < SECTION_OUT_OF_RANGE);
 	bool anyLinearRecordingAfter = false;
 
 	// Now action the launching of Clips
@@ -634,7 +645,8 @@ doNormalLaunch:
 
 		// If we found a playing Clip outside of the armed section, or vice versa, then we can't say we legitimately
 		// just launched a section
-		if (clip->launchStyle != LaunchStyle::FILL && clipActiveAfter != (clip->section == lastSectionArmed)) {
+		if (clip->launchStyle != LaunchStyle::FILL && clipActiveAfter != (clip->section == lastSectionArmed)
+		    && currentSong->sections[clip->section].numRepetitions != LAUNCH_EXCLUSIVE) {
 			sectionWasJustLaunched = false;
 		}
 	}
@@ -650,9 +662,9 @@ doNormalLaunch:
 		// Otherwise...
 		else {
 
-			bool sectionManuallyStopped = (lastSectionArmed == 254);
+			bool sectionManuallyStopped = (lastSectionArmed == SECTION_OUT_OF_RANGE);
 
-			lastSectionArmed = 255;
+			lastSectionArmed = REALLY_OUT_OF_RANGE;
 
 			// If no Clips active anymore...
 			if (!anyClipsStillActiveAfter) {
@@ -674,7 +686,7 @@ doNormalLaunch:
 
 						int32_t sectionToArm;
 						// And re-activate the first section
-						if (currentSong->sectionToReturnToAfterSongEnd < 254) {
+						if (currentSong->sectionToReturnToAfterSongEnd < SECTION_OUT_OF_RANGE) {
 							sectionToArm = currentSong->sectionToReturnToAfterSongEnd;
 						}
 						else {
@@ -976,7 +988,9 @@ void Session::toggleClipStatus(Clip* clip, int32_t* clipIndex, bool doInstant, i
 		return;
 	}
 
-	lastSectionArmed = 255;
+	if (currentSong->sections[clip->section].numRepetitions != LAUNCH_EXCLUSIVE) {
+		lastSectionArmed = REALLY_OUT_OF_RANGE;
+	}
 
 	// If Clip armed, cancel arming - but not if it's an "instant" toggle
 	if (clip->launchStyle == LaunchStyle::FILL && clip->armState != ArmState::OFF && !clip->isActiveOnOutput()) {
@@ -1138,7 +1152,7 @@ void Session::armClipToStopAction(Clip* clip) {
 }
 
 void Session::soloClipAction(Clip* clip, int32_t buttonPressLatency) {
-	lastSectionArmed = 255;
+	lastSectionArmed = REALLY_OUT_OF_RANGE;
 
 	bool anyClipsDeleted = false;
 
@@ -1188,7 +1202,7 @@ renderAndGetOut:
 void Session::armSection(uint8_t section, int32_t buttonPressLatency) {
 
 	// Get rid of soloing. And if we're not a "share" section, get rid of arming too
-	currentSong->turnSoloingIntoJustPlaying(currentSong->sections[section].numRepetitions != -1);
+	currentSong->turnSoloingIntoJustPlaying(currentSong->sections[section].numRepetitions >= 0);
 
 	// If every Clip in this section is already playing, and no other Clips are (unless we're a "share" section), then
 	// there's no need to launch the section because it's already playing. So, make sure this isn't the case before we
@@ -1202,7 +1216,7 @@ void Session::armSection(uint8_t section, int32_t buttonPressLatency) {
 		}
 
 		// If a Clip in another section is playing and we're not a "share" section...
-		if (currentSong->sections[section].numRepetitions != -1 && clip->section != section
+		if (currentSong->sections[section].numRepetitions > -1 && clip->section != section
 		    && ((clip->armState != ArmState::OFF) != clip->activeIfNoSolo)) {
 			goto yupThatsFine;
 		}
@@ -1235,7 +1249,10 @@ yupThatsFine:
 		userWantsToArmClipsToStartOrSolo(
 		    section, nullptr, stopAllOtherClips, false,
 		    false); // Don't allow "late start". It's too fiddly to implement, and rarely even useful for sections
-		lastSectionArmed = section;
+
+		if (currentSong->sections[section].numRepetitions != LAUNCH_EXCLUSIVE) {
+			lastSectionArmed = section;
+		}
 	}
 
 	armingChanged();
@@ -1255,7 +1272,8 @@ void Session::armSectionWhenNeitherClockActive(ModelStack* modelStack, int32_t s
 			modelStack->song->assertActiveness(modelStackWithTimelineCounter);
 		}
 
-		if (stopAllOtherClips && clip->section != section && clip->activeIfNoSolo) {
+		if (stopAllOtherClips && clip->section != section && clip->activeIfNoSolo
+		    && currentSong->sections[clip->section].numRepetitions != LAUNCH_EXCLUSIVE) {
 			// thisClip->expectNoFurtherTicks(currentSong); // No, don't need this, cos it's not playing!
 			clip->activeIfNoSolo = false;
 		}
@@ -1689,8 +1707,15 @@ wantActive:
 				if (stopAllOtherClips) {
 
 weWantThisClipInactive:
+					// Clip is launched exclusively
+					if (currentSong->sections[thisClip->section].numRepetitions == LAUNCH_EXCLUSIVE) {
+						if (thisClip->activeIfNoSolo) {
+							thisClip->armState = ArmState::OFF;
+						}
+					}
+
 					// If it's active, arm it to stop
-					if (thisClip->activeIfNoSolo) {
+					else if (thisClip->activeIfNoSolo) {
 						thisClip->armState = ArmState::ON_NORMAL;
 					}
 
@@ -1941,8 +1966,8 @@ void Session::armClipLowLevel(Clip* clipToArm, ArmState armState, bool mustUnarm
 int32_t Session::userWantsToArmNextSection(int32_t numRepetitions) {
 
 	int32_t currentSection = getCurrentSection();
-	if (currentSection < 254) {
-		if (numRepetitions == -1) {
+	if (currentSection < SECTION_OUT_OF_RANGE) {
+		if (numRepetitions <= -1) {
 			numRepetitions = currentSong->sections[currentSection].numRepetitions;
 		}
 
@@ -1960,10 +1985,10 @@ int32_t Session::userWantsToArmNextSection(int32_t numRepetitions) {
 int32_t Session::getCurrentSection() {
 
 	if (currentSong->getAnyClipsSoloing()) {
-		return 255;
+		return REALLY_OUT_OF_RANGE;
 	}
 
-	int32_t section = 255;
+	int32_t section = REALLY_OUT_OF_RANGE;
 
 	bool anyUnlaunchedLoopablesInSection[kMaxNumSections];
 	memset(anyUnlaunchedLoopablesInSection, 0, sizeof(anyUnlaunchedLoopablesInSection));
@@ -1971,12 +1996,17 @@ int32_t Session::getCurrentSection() {
 	for (int32_t l = 0; l < currentSong->sessionClips.getNumElements(); l++) {
 		Clip* clip = currentSong->sessionClips.getClipAtIndex(l);
 
+		// Launch exclusively section
+		if (currentSong->sections[clip->section].numRepetitions == LAUNCH_EXCLUSIVE) {
+			continue;
+		}
+
 		if (clip->activeIfNoSolo) {
-			if (section == 255) {
+			if (section == REALLY_OUT_OF_RANGE) {
 				section = clip->section;
 			}
 			else if (section != clip->section) {
-				return 254;
+				return SECTION_OUT_OF_RANGE;
 			}
 		}
 		else {
@@ -1988,7 +2018,7 @@ int32_t Session::getCurrentSection() {
 	}
 
 	if (anyUnlaunchedLoopablesInSection[section]) {
-		return 255;
+		return REALLY_OUT_OF_RANGE;
 	}
 	return section;
 }
@@ -2031,12 +2061,12 @@ void Session::setupPlayback() {
 
 	currentSong->setParamsInAutomationMode(playbackHandler.recording == RecordingMode::ARRANGEMENT);
 
-	lastSectionArmed = 255;
+	lastSectionArmed = REALLY_OUT_OF_RANGE;
 }
 
 // Returns whether to do an instant song swap
 bool Session::endPlayback() {
-	lastSectionArmed = 255;
+	lastSectionArmed = REALLY_OUT_OF_RANGE;
 
 	bool anyClipsRemoved = currentSong->deletePendingOverdubs();
 
@@ -2260,7 +2290,7 @@ traverseClips:
 			// If we're doing a song swap...
 			if (preLoadedSong) {
 				cancelAllLaunchScheduling();
-				lastSectionArmed = 255;
+				lastSectionArmed = REALLY_OUT_OF_RANGE;
 				playbackHandler.doSongSwap();
 				swappedSong = true;
 
@@ -2323,11 +2353,11 @@ traverseClips:
 	if (playbackHandler.lastSwungTickActioned == 0 || enforceSettingUpArming) {
 		int32_t currentSection = userWantsToArmNextSection();
 
-		if (currentSection < 254 && currentSong->areAllClipsInSectionPlaying(currentSection)) {
+		if (currentSection < SECTION_OUT_OF_RANGE && currentSong->areAllClipsInSectionPlaying(currentSection)) {
 			currentSong->sectionToReturnToAfterSongEnd = currentSection;
 		}
 		else {
-			currentSong->sectionToReturnToAfterSongEnd = 255;
+			currentSong->sectionToReturnToAfterSongEnd = REALLY_OUT_OF_RANGE;
 		}
 	}
 
@@ -2584,7 +2614,7 @@ void Session::stopOutputRecordingAtLoopEnd() {
 	// If no launch-event currently, plan one
 	if (!launchEventAtSwungTickCount) {
 		armAllClipsToStop(1);
-		lastSectionArmed = 254;
+		lastSectionArmed = SECTION_OUT_OF_RANGE;
 		armingChanged();
 	}