+++ CHANGELOG.md
@@ -9,8 +9,13 @@
 - Added a boss/roland style Dimension effect
 - Made grain much faster and updated controls
 - Added LPF to Mutable Instruments Reverb
-- Added an option to reverse the stutter effect
 - Added two more envelopes (Envelope 3 and Envelope 4), which you can access from the sound editor menu.
+
+#### <ins>Per-clip Stutter with options: Quantize, Reverse, and Ping-Pong</ins>
+- Now the stutter buffer can also be set to play in reverse, and ping-pong.
+- A new menu has been added to the `SONG` menu, under `FX` submenu, to set the stutter configuration for the song.
+- An extra menu has been added per sound source (either synths, kit rows, affect-entire kits, or audio clips) to
+set the stutter configuration independently of the song configuration if you wish.
 
 ### User Interface
 
@@ -154,6 +159,7 @@ also affect normal sequenced notes while arpeggiator is Off.
 - Added `CHORD PROBABILITY (CHRD)` paramater that will allow you to control the chance of a note to play, at the same time, itself and other notes to form a chord (5th, 3rd and 7th).
 - Added `CHORD POLYPHONY (POLY)` paramater that works together with the previous parameter, and this one will dictate how big is the chord, that is, how many notes will be played at the same time.
 - Added `LOCK` parameter to allow you to freeze the current set of randomized values so the sequence has a repeatable pattern.
+- Added new `WALK1`, `WALK2`, `WALK3` and `PATTERN` note modes.
 
 ### MIDI
 
@@ -252,6 +258,7 @@ also affect normal sequenced notes while arpeggiator is Off.
 - A white playhead is now rendered in Song Grid and Performance Views that let's you know when a clip or section launch event is scheduled to occur. The playhead only renders the last 16 notes before a launch event.
   - Note: this playhead can be turned off in the Community Features submenu titled: `Enable Launch Event Playhead (PLAY)`
 - The display now shows the number of Bars (or Quarter Notes for the last bar) remaining until a clip or section launch event in all Song views (Grid, Row, Performance).
+- A new option, 'Launch Exclusively', isolates a clip section from all other launch activity. This option is found to the left of option 'Launch non-exclusively' when selecting the section's number of repetitions. As a complement to non-exclusive sections that arm and turn off when another section is launched, exclusive sections remain independant.
 
 #### <ins>Audio Clips</ins>